Basic terrain generation

+++ src/main/scala/org/kolinek/gengame/geometry/Triangle.scala
@@ -2,6 +2,6 @@ package org.kolinek.gengame.geometry
 
 import org.kolinek.gengame.terragen.mesh.NormalCalculator
 
-case class Triangle(a: Long, b: Long, c: Long) {
+case class Triangle(a: Int, b: Int, c: Int) {
     def toList = List(a, b, c)
 }
